improving

+++ customize.sh
@@ -14,13 +14,15 @@ set_perm $MODPATH/system/bin/znull 0 0 0755
 set_perm $MODPATH/system/bin/adaway 0 0 0755
 set_perm $MODPATH/system/bin/adbport 0 0 0755
 
-echo -e "\nzNull Project desu...\n"
+echo
+cat $MODPATH/libraries/zNull-Project.txt
 
-mv -f $MODPATH/libraries/znull.conf /data/media/0/Android/. && echo -e "You can change package name on Internal/Android/znull.conf\nor use znull [-s|set].\n"
+
+if [[ ! -s /data/media/0/Android/znull.conf ]];then
+	mv -f $MODPATH/libraries/znull.conf /data/media/0/Android/. && echo -e "You can change package name on Internal/Android/znull.conf\nor use znull [-s|set].\n"
+fi
 
 mv -f $MODPATH/libraries/zNull-SCRIPT.lua /data/media/0/. && echo -e "You can use our script on Internal.\nOur script is OnLy for PUBG MOBILE LITE and GLOBAL!"
-
-
 rm -rf $MODPATH/libraries/*.lua
 
 echo -e "Usage: su -c znull [-h|help|--help]"
